+++ src/mongo/db/ops/modifier_pull_all.cpp
@@ -40,7 +40,7 @@ namespace mongo {
         mutablebson::Document& doc;
 
         // Index in _fieldRef for which an Element exist in the document.
-        int32_t pathFoundIndex;
+        size_t pathFoundIndex;
 
         // Element corresponding to _fieldRef[0.._idxFound].
         mutablebson::Element pathFoundElement;
@@ -197,8 +197,7 @@ namespace mongo {
         mutablebson::Document& doc = logBuilder->getDocument();
 
         const bool pathExists = _preparedState->pathFoundElement.ok() &&
-            (_preparedState->pathFoundIndex ==
-             static_cast<int32_t>(_fieldRef.numParts() - 1));
+            (_preparedState->pathFoundIndex == (_fieldRef.numParts() - 1));
 
         // value for the logElement ("field.path.name": <value>)
         mutablebson::Element logElement = pathExists ?
